Auth (#75)
* adding firebase

* Using .tom file for plugins

* Sketch of auth

* Successfully creating an user.

The UI looks like sh** at the moment

* Login and logout working

* Adding show password button

* adding better auth menu

* Adding name for user

* Fixing start of navigation

* inner navigation

* Fixing activity usage in the auth process

* Fixing new document creation

* adding aws amplify

* Adding auth

* sign in working

* Creating module

* Navigation debug

* Moving code

* Doing logout

* Auth menu design

* Making AuthMenu look better

* Adding new auth

* Adding number to StoryTypes for performance

* Fixing load of notes

* Filtering

* Fixing filter logic

* Adding more options

* Create file to share

* removing unused field in StoryStep

* Serialization configured

* Sending content as text

* Creating IntroNotes inside android project

* Creating sketch of modules

* Moving models to common module

* Post to write write notes working

* Fixing IntroNotes lambda

* Adding network module

* Sketch of network module

* Deleting unnecessary class

* Using Instant instead of Date

* Updating lastUpdatedAt correctly

* Remove case sensitivity in order

* Adding serialization

* First request working!

* Making connected intro note look a better

* test dynamodb locally

* deleting scripts

+++ application/editor/src/main/java/com/github/leandroborgesferreira/storytellerapp/editor/NoteEditorViewModel.kt
@@ -1,17 +1,26 @@
 package com.github.leandroborgesferreira.storytellerapp.editor
 
-import androidx.lifecycle.ViewModel
+import android.app.Application
+import android.content.Context
+import android.content.Intent
+import androidx.lifecycle.AndroidViewModel
 import androidx.lifecycle.viewModelScope
-import com.github.leandroborgesferreira.storytellerapp.editor.model.EditState
 import com.github.leandroborgesferreira.storyteller.backstack.BackstackHandler
 import com.github.leandroborgesferreira.storyteller.backstack.BackstackInform
+import com.github.leandroborgesferreira.storyteller.filter.DocumentFilter
+import com.github.leandroborgesferreira.storyteller.filter.DocumentFilterObject
 import com.github.leandroborgesferreira.storyteller.manager.DocumentRepository
 import com.github.leandroborgesferreira.storyteller.manager.StoryTellerManager
 import com.github.leandroborgesferreira.storyteller.model.action.Action
-import com.github.leandroborgesferreira.storyteller.model.story.Decoration
 import com.github.leandroborgesferreira.storyteller.model.story.DrawState
 import com.github.leandroborgesferreira.storyteller.model.story.StoryState
+import com.github.leandroborgesferreira.storyteller.models.document.Document
+import com.github.leandroborgesferreira.storyteller.models.story.Decoration
+import com.github.leandroborgesferreira.storyteller.serialization.extensions.toApi
+import com.github.leandroborgesferreira.storyteller.serialization.json.storyTellerJson
+import com.github.leandroborgesferreira.storyteller.serialization.request.wrapInRequest
 import com.github.leandroborgesferreira.storyteller.utils.extensions.noContent
+import com.github.leandroborgesferreira.storytellerapp.editor.model.EditState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.SharingStarted
@@ -21,16 +30,23 @@ import kotlinx.coroutines.flow.map
 import kotlinx.coroutines.flow.stateIn
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.withContext
+import kotlinx.serialization.encodeToString
+
 
 class NoteEditorViewModel(
     val storyTellerManager: StoryTellerManager,
-    private val documentRepository: DocumentRepository
-) : ViewModel(),
+    private val documentRepository: DocumentRepository,
+    private val documentFilter: DocumentFilter = DocumentFilterObject,
+    application: Application
+) : AndroidViewModel(application),
     BackstackInform by storyTellerManager,
     BackstackHandler by storyTellerManager {
 
     private val _editModeState = MutableStateFlow(true)
     val editModeState: StateFlow<Boolean> = _editModeState
+
+    private val _showGlobalMenu = MutableStateFlow(false)
+    val showGlobalMenu = _showGlobalMenu.asStateFlow()
 
     private val _editHeader = MutableStateFlow(false)
     val editHeader = _editHeader.asStateFlow()
@@ -65,10 +81,10 @@ class NoteEditorViewModel(
         storyTellerManager.newStory(documentId, title)
 
         viewModelScope.launch {
-            val document = storyTellerManager.currentDocument.stateIn(this).value
-
-            documentRepository.saveDocument(document)
-            storyTellerManager.saveOnStoryChanges(documentRepository)
+            storyTellerManager.currentDocument.stateIn(this).value?.let { document ->
+                documentRepository.saveDocument(document)
+                storyTellerManager.saveOnStoryChanges()
+            }
         }
     }
 
@@ -80,7 +96,7 @@ class NoteEditorViewModel(
 
             if (document != null) {
                 storyTellerManager.initDocument(document)
-                storyTellerManager.saveOnStoryChanges(documentRepository)
+                storyTellerManager.saveOnStoryChanges()
             }
         }
     }
@@ -89,7 +105,7 @@ class NoteEditorViewModel(
         viewModelScope.launch(Dispatchers.IO) {
             val document = storyTellerManager.currentDocument.stateIn(this).value
 
-            if (story.value.stories.noContent()) {
+            if (document != null && story.value.stories.noContent()) {
                 documentRepository.deleteDocument(document)
 
                 withContext(Dispatchers.Main) {
@@ -120,6 +136,47 @@ class NoteEditorViewModel(
 
     fun onHeaderEditionCancel() {
         _editHeader.value = false
+    }
+
+    fun onMoreOptionsClick() {
+        _showGlobalMenu.value = !_showGlobalMenu.value
+    }
+
+    fun shareDocumentInJson(context: Context) {
+        val json = storyTellerJson
+
+        viewModelScope.launch {
+            val document: Document = storyTellerManager.currentDocument
+                .stateIn(viewModelScope)
+                .value ?: return@launch
+
+            val documentTitle = document.title.replace(" ", "_")
+
+            val apiContent = documentFilter.removeMetaData(document.content)
+                .map { (position, story) ->
+                    story.toApi(position)
+                }
+
+            val request = document.toApi().copy(content = apiContent).wrapInRequest()
+            val jsonDocument = json.encodeToString(request)
+
+            val intent = Intent(Intent.ACTION_SEND).apply {
+                addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+                putExtra(Intent.EXTRA_TEXT, jsonDocument)
+                putExtra(Intent.EXTRA_TITLE, documentTitle)
+                action = Intent.ACTION_SEND
+                type = "application/json"
+            }
+
+            context.startActivity(
+                Intent.createChooser(intent, "Export Document")
+            )
+        }
+    }
+
+    override fun onCleared() {
+        super.onCleared()
+        storyTellerManager.onClear()
     }
 }
